+++ R/excess_attenuation.R
@@ -1,7 +1,7 @@
 #' Measure excess attenuation
 #' 
 #' \code{excess_attenuation} measures excess attenuation in signals referenced in an extended selection table.
-#' @usage excess_attenuation(X, parallel = 1, pb = TRUE, method = 1, 
+#' @usage excess_attenuation(X, parallel = 1, pb = TRUE, method = 1, type = "Marten",
 #' bp = NULL, output = "est", hop.size = 1, wl = NULL, ovlp = 70)
 #' @param X object of class 'extended_selection_table' created by the function \code{\link[warbleR]{selection_table}} from the warbleR package. The data frame must include the following additional columns: 'distance', 'signal.type', 'bottom.freq' and 'top.freq'.
 #' @param parallel Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing).
@@ -10,6 +10,11 @@
 #' \itemize{
 #' \item \code{1}: compare all signals with their counterpart that was recorded at the closest distance to source (e.g. compare a signal recorded at 5m, 10m and 15m with its counterpart recorded at 1m). This is the default method. 
 #' \item \code{2}: compare all signals with their counterpart recorded at the distance immediately before (e.g. a signal recorded at 10m compared with the same signal recorded at 5m, then signal recorded at 15m compared with same signal recorded at 10m and so on).
+#' }
+#' @param type Character vector of length 1 to indicate the 'type' of excess attenuation to be used. Two types are available:
+#' \itemize{
+#' \item \code{Marten}: as described by Marten et al. 1977:(total_attenuation - spheric_spreading_attenuation) / distance. This is the default method. Attenuation is measured as changes in energy on amplitude RMS (root mean square).
+#' \item \code{Darden}: as described by Darden et al 2008: microphone_gain - 20 x log(distance / 10) - 20 x log(envelope_correlation). The function \code{\link{envelope_correlation}} is used internally. Microphone gain is the combined microphone gain of the reference and re-recorded signals.
 #' }
 #' @param bp Numeric vector of length 2 giving the lower and upper limits of a frequency bandpass filter (in kHz). Default is \code{NULL}.
 #' @param output Character vector of length 1 to determine if an extended selection table ('est', default) or a data frame ('data.frame').
@@ -38,17 +43,19 @@
 #' }
 #' 
 #' @author Marcelo Araya-Salas (\email{marceloa27@@gmail.com})
-#' @seealso \code{\link{spcc}}
+#' @seealso \code{\link{spcc}}; \code{\link{envelope_correlation}}
 #' @references {
-#' Araya-Salas, M. (2020),=. baRulho: baRulho: quantifying habitat-induced degradation of (animal) acoustic signals in R. R package version 1.0.0
+#' Araya-Salas, M. (2020). baRulho: baRulho: quantifying habitat-induced degradation of (animal) acoustic signals in R. R package version 1.0.0
 #' 
-#' Marten, K., & Marler, P. (1977). Sound transmission and its significance for animal vocalization. Behavioral Ecology and Sociobiology, 2(3), 271-290.
+#' Darden, SK, Pedersen SB, Larsen ON, & Dabelsteen T. (2008). Sound transmission at ground level in a short-grass prairie habitat and its implications for long-range communication in the swift fox *Vulpes velox*. The Journal of the Acoustical Society of America, 124(2), 758-766.
 #' 
-#' Morton, E. S. (1975). Ecological sources of selection on avian sounds. The American Naturalist, 109(965), 17-34.
+#' Marten K, & Marler P. (1977). Sound transmission and its significance for animal vocalization. Behavioral Ecology and Sociobiology, 2(3), 271-290.
+#' 
+#' Morton ES. (1975). Ecological sources of selection on avian sounds. The American Naturalist, 109(965), 17-34.
 #' }
 #last modification on nov-01-2019 (MAS)
 
-excess_attenuation <- function(X, parallel = 1, pb = TRUE, method = 1, 
+excess_attenuation <- function(X, parallel = 1, pb = TRUE, method = 1, type = "Marten", 
                                bp = NULL, output = "est", hop.size = 1, wl = NULL, ovlp = 70){
   
   # set pb options 
@@ -98,7 +105,7 @@ excess_attenuation <- function(X, parallel = 1, pb = TRUE, method = 1,
     
     # get RMS for signal
     sigRMS <- seewave::rms(seewave::env(clp, f = clp@samp.rate, envt = "abs", plot = FALSE))
-    sigRMS <- 20 * log10(sigRMS)
+    sigRMS <- - 20 * log10(sigRMS)
     
     return(data.frame(X[y, , drop = FALSE], sigRMS))
   }
@@ -120,7 +127,7 @@ excess_attenuation <- function(X, parallel = 1, pb = TRUE, method = 1,
   RMS_list <- split(RMS_df, RMS_df$signal.type)
   
   # calculate excess attenuation
-  X_list <- sapply(RMS_list, function(Y, meth = method){
+  X_list <- sapply(RMS_list, function(Y, meth = method, tp = type){
     
     if (Y$signal.type[1] == "ambient") Y$excess.attenuation <- NA else {
     
@@ -131,9 +138,11 @@ excess_attenuation <- function(X, parallel = 1, pb = TRUE, method = 1,
       sig_RMS_REF <- Y$sigRMS[which.min(Y$distance)]
       dist_REF <- Y$distance[which.min(Y$distance)]
       
+      # type Marten
+      if (tp == "Marten"){
       # term 1: decrease in signal amplitude (RMS) of reference (Ref) vs re-recorded (RR)
       term1 <- sig_RMS_REF - Y$sigRMS 
-      
+        
       # lost due to spheric spreading
       term2 <- -20 * log10(1 / Y$distance)
       
@@ -141,7 +150,28 @@ excess_attenuation <- function(X, parallel = 1, pb = TRUE, method = 1,
       term3 <- Y$distance - dist_REF
       
       # excess attenuation = (total attenuation - spheric spreading attenuation) / distance
-      Y$excess.attenuation <- (term1 + term2) / term3
+      ea <- (term1 + term2) / term3
+      } 
+      
+      if (tp == "Darden"){
+        
+        #EA = g - 20 log(d / 10) - 20 log(k)
+        # term1 = g (combined mic gain)
+        term1 <- sapply(Y$sigRMS, function(x) seewave::moredB(c(sig_RMS_REF, x)), USE.NAMES = FALSE)
+        
+        # term2 = - 20 log(d / 10)
+        term2 <- - 20 * log10(Y$distance / 10)
+        
+        # term3 = - 20 log(k)
+        # get envelope correlation (k)
+       k <- envelope_correlation(X[X$signal.type == Y$signal.type[1],], output = "data.frame", pb = FALSE)$envelope.correlation
+        term3 <- -20 * log10(k)
+        
+        # excess attenuation = (total attenuation - spheric spreading attenuation) / distance
+        ea <- term1 + term2 + term3
+      } 
+      
+      Y$excess.attenuation <- ea
       Y$excess.attenuation[which.min(Y$distance)] <- NA
       }
     
@@ -154,7 +184,9 @@ excess_attenuation <- function(X, parallel = 1, pb = TRUE, method = 1,
       # sort by distance
       Y <- Y[order(Y$distance), ]
       
-      # term 1: decrease in signal amplitude (RMS) of reference (Ref) vs re-recorded (RR)
+      
+      if (tp == "Marten"){
+        # term 1: decrease in signal amplitude (RMS) of reference (Ref) vs re-recorded (RR)
       term1 <- Y$sigRMS[-nrow(Y)] - Y$sigRMS[-1] 
       
       # lost due to spheric spreading
@@ -164,11 +196,38 @@ excess_attenuation <- function(X, parallel = 1, pb = TRUE, method = 1,
       term3 <- Y$distance[-1] - Y$distance[-nrow(Y)]
       
       # excess attenuation = (total attenuation - spheric spreading attenuation) / distance
-      EA <- (term1 + term2) / term3
+      ea <- (term1 + term2) / term3
       
       # add NA for first distance
-      Y$excess.attenuation <- c(NA, EA)      
+      ea <- c(NA, ea)      
+      }
       
+      # type Darden
+      if (tp == "Darden"){
+        
+        #EA = g - 20 log(d / 10) - 20 log(k)
+        # term1 = g (combined mic gain)
+        term1 <- sapply(2:nrow(Y), function(x) seewave::moredB(c(Y$sigRMS[x - 1], Y$sigRMS[x])), USE.NAMES = FALSE)
+        
+        # term2 = - 20 log(d / 10)
+        term2 <- - 20 * log10(Y$distance[-1] - Y$distance[-nrow(Y)] / 10)
+        
+        # term3 = - 20 log(k)
+        # get envelope correlation (k)
+        k <- envelope_correlation(X[X$signal.type == Y$signal.type[1],], output = "data.frame", pb = FALSE, method = 2)$envelope.correlation
+        
+        # order by distance too
+        k <- k[order(X$distance[X$signal.type == Y$signal.type[1]])]
+        term3 <- -20 * log10(k[-1])
+        
+        # excess attenuation
+        ea <- term1 + term2 + term3
+        
+        # add NA for first distance
+        ea <- c(NA, ea)  
+      } 
+      
+      Y$excess.attenuation <- ea
       # reorder results
       Y <- Y[order(Y$org....ord), ]